Init scan view created with grid systems and empty functions
